requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.8.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
colorama==0.4.3
crayons==0.4.0
crispy-tailwind==0.2.0
dateparser==1.0.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-advanced-reports==0.9.24
django-celery==3.3.1
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.3.0
django-report-builder==6.3.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
docutils==0.16
env==0.1.0
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
halo==0.0.30
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==3.0.37
lockfile==0.12.2
log-symbols==0.0.14
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mock==4.0.2
mysqlclient==2.0.1
node==0.9.25
note==0.5.2
npm==0.1.1
odfpy==1.4.1
odict==1.7.0
openpyxl==3.0.5
optional-django==0.3.0
packaging==20.4
pexpect==4.8.0
Pillow==7.2.0
pluggy==0.13.1
plumber==1.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.9.0
Pygments==2.7.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.1
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
react==4.3.0
react-pages==0.3.2
regex==2020.11.13
requests==2.24.0
selenium==3.141.0
six==1.15.0
spinners==0.0.24
sqlparse==0.3.1
Strip==0.0.1.dev14
stripe==2.55.1
tablib==2.0.0
termcolor==1.1.0
toml==0.10.1
tornado==6.1
tzlocal==2.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.1.0
Werkzeug==1.0.1
wfastcgi==3.0.0
whitenoise==5.2.0
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zope.component==4.6.2
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.2.0
zope.lifecycleevent==4.3
zope.proxy==4.3.5
